+++ lib/jsdom/living/helpers/style-rules.js
@@ -1,5 +1,7 @@
 "use strict";
 const cssom = require("cssom");
+const { CustomPropertyDependencyGraph,
+        isCustomProperty } = require("./custom-property-dependency-graph");
 const defaultStyleSheet = require("../../browser/default-stylesheet");
 const { matchesDontThrow } = require("./selectors");
 const { propertiesWithResolvedValueImplemented } = require("./resolved-properties");
@@ -99,92 +101,8 @@ function getSpecifiedValue(element, property) {
   return getDefaultValue(element, property);
 }
 
-// // https://www.w3.org/TR/css-variables-1/#using-variables
-// function _getIndividualVarSubstitute(element, property, variable) {
-//   const matches = typeof variable !== 'string' ? [] :
-//         variable.matchAll(/^var\((--[-\w]+),?\s*(.*)\)/g);
-
-//   for (let match of matches) {
-//     const [_, customProp, fallback] = [...match];
-
-//     let substitute = exports.getResolvedValue(element, customProp);
-//     if (!substitute.value && fallback) {
-//       let fallbackValMatches = [...fallback.matchAll(/([^,()]+(?:\(.*?\))?)/g)];
-//       const fallbackVals = fallbackValMatches.map(match => {
-//         const [_, arg] = [...match];
-//         if (arg === 'initial' || arg === 'inherit' ||
-//             arg === 'unset' || arg === 'revert') {
-//           return getDefaultValue(element, property, arg);
-//         }
-        
-//         const fallbackVal = {
-//           value: arg,
-//           priority: null,
-//         };
-//         return getVarSubstitute(element, property, fallbackVal);
-//       });
-
-//       substitute.value = fallbackVals.map(e=>e.value).join(',');
-//     }
-
-//     // TODO: Check for animation-taintedness
-//     // https://www.w3.org/TR/css-variables/#animation-tainted
-//     //
-//     // Left this undone since there's currently no way to check whether a
-//     // property is animatable. 
-
-//     return substitute.value;
-//   }
-
-//   return '';
-// }
-
-// function getCssVars(propertyValue) {
-//   const cssVars = [];
-//   let buffer = '';
-//   let startOfCurrentVar = undefined;
-//   let openParens = 0;
-
-//   for (let ii = 0; ii < propertyValue.length; ii++) {
-//     const newChar = propertyValue[ii];
-//     buffer += newChar;
-//     if (startOfCurrentVar === undefined && buffer.includes('var(')) {
-//       buffer = 'var(';
-//       startOfCurrentVar = ii - buffer.length - 1;
-//       openParens = 1; 
-//     } else if (startOfCurrentVar !== undefined && newChar === '(') {
-//       openParens += 1;
-//     } else if (startOfCurrentVar !== undefined && newChar === ')') {
-//       openParens -= 1;
-//       if (openParens === 0) {
-//         cssVars.push(buffer);
-//         startOfCurrentVar = undefined;
-//         buffer = '';
-//       }
-//     }
-//   }
-
-//   return cssVars;
-// }
-
-// function _getVarSubstitute(element, propertyName, propertyInfo) {
-//   const cssVars = getCssVars(propertyInfo.value);
-
-//   const substitute = cssVars.map(v => {
-//     return _getIndividualVarSubstitute(element, propertyName, v);
-//   }).join(' ');
-                                
-//   return {
-//     value: substitute ? substitute : propertyInfo.value,
-//     priority: propertyInfo.priority,
-//   };
-// }
-
-function isCustomProperty(propertyName) {
-  return typeof propertyName === 'string' &&
-    propertyName.indexOf('--') === 0;
-}
-
+// Parses a single var() from the property value. Returns
+// a list of [precedingText, cssVar, followingText]
 function parseCssVar(propertyValue) {
   let buffer = '';
   let openParens = 0;
@@ -214,7 +132,9 @@ function parseCssVar(propertyValue) {
   return [null, null, null];
 }
 
+// https://www.w3.org/TR/css-variables-1/#syntax
 function isCssVarInvalid(leadingText, cssVar, trailingText) {
+  
   // Invalidate if immediately followed by CSS token
   const tokenRegex = /^\s*(?:in|cm|em|mm|pt|pc|px|ex|rem|vh|vw|ch|deg|grad|rad|%)/;
   if (trailingText && tokenRegex.test(trailingText)) {
@@ -224,6 +144,10 @@ function isCssVarInvalid(leadingText, cssVar, trailingText) {
   return false;
 }
 
+// https://www.w3.org/TR/css-variables-1/#using-variables
+//
+// Specifically, this handles a single variable and its fallback
+// Expected input: var(--a, red)
 function getSingleCssVarSubstitute(element, cssVar) {
   const cssVarArgs = cssVar.match(
     /^var\((?<customProp>--[-\w]+),?\s*(?<fallback>.*)\)/);
@@ -245,7 +169,11 @@ function getSingleCssVarSubstitute(element, cssVar) {
   return guaranteedInvalid;
 }
 
-function normalizeSubstitution(substitute) {
+// Modifies the substitution string based on the surrounding chars.
+//
+// At the moment, specific to the case where two var()s are back-to-back
+// like: var(--a)var(--b) => var(--a) var(--b)
+function normalizeCssVarSubstitution(substitute) {
   return (match, offset, string) => {
     // Add leading space if immediately against something else
     if (offset > 0 && /[^,\s(]/.test(string[offset-1])) {
@@ -256,6 +184,13 @@ function normalizeSubstitution(substitute) {
   };
 }
 
+// https://www.w3.org/TR/css-variables-1/#using-variables
+//
+// Specifically, this function parses a specified value of (possibly)
+// many var()s and their fallbacks, substituting the computed value
+// for each var().
+//
+// Expected input: var(--a, red) var(--b, var(--c, var(--d)))
 function makeCssVarSubstitutions(element, computedPropertyInfo) {
   let cssVar;
   const substitute = computedPropertyInfo;
@@ -272,7 +207,7 @@ function makeCssVarSubstitutions(element, computedPropertyInfo) {
       const cssVarSubstitute = getSingleCssVarSubstitute(element, cssVar);
       substitute.value = substitute.value.replace(
         cssVar,
-        normalizeSubstitution(cssVarSubstitute));
+        normalizeCssVarSubstitution(cssVarSubstitute));
         
       substitute.priority = (
         cssVarSubstitute.priority || substitute.priority ?
@@ -290,201 +225,25 @@ function makeCssVarSubstitutions(element, computedPropertyInfo) {
   return substitute;
 }
 
-// function getVarListSubstitute(element, cssVarList) {
-//   let substitutes = [];
-//   for (let cssVar of cssVarList) {
-//     const cssVarArgs = cssVar.matchAll(/^var\((--[-\w]+),?\s*(.*)\)/g);
-
-//     for (let cssVarArg of cssVarArgs) {
-//       const [_, customProp, fallback] = [...cssVarArg];
-//       let substitute = getVarSubstitute(element, customProp);
-//       if (!substitute.value) {
-//         const fallbackCssVars = getCssVars(fallback);
-//         if (fallbackCssVars.length === 0) {
-//           substitute.value = fallback;
-          
-//         } else {
-//           substitute = getVarListSubstitute(element, fallbackCssVars);
-//         }
-//       }
-      
-//       substitutes.push(substitute);
-//     }
-//   }
-
-//   return {
-//     value: substitutes.map(e => e.value).join(' '),
-//     priority: substitutes.some(e => e.priority !== null) ? 'important' : null,
-//   };
-// }
-
-class CustomPropertyGraph {
-  constructor(element) {
-    this._element = element;
-    this._nodes = [];
-  }
-
-  createNodesFrom(propertyName) {
-    // Build graph
-    //   propertyName is first node in graph
-    //   1. for each custom property in propertyInfo (use regex 'var(--[\w]+')
-    //        create node
-    //        draw outgoing edge to each new node (find using getCascadedPropertyValue)
-    //        for each new node
-    //          if the node is a custom property that does not have outgoing edges 
-    //            return to 1
-    this._nodes = [];
-    this._attachNodes(undefined, propertyName);
-  }
-
-  isCyclicAround(propertyName) {
-    // Check for cycles
-    //   start on propertyName
-    //   2. Mark node as visited
-    //        for each outgoing edge,
-    //          if node on other end is not visited,
-    //            return to 2
-    //          if node on other end is (original) propertyName
-    //            return true
-    const node = this._getNode(propertyName);
-    const isCyclic = node.find(propertyName);
-    this._unvisitNodes();
-    return isCyclic;
-  }
-
-  findSubstitute(propertyName) {
-    // Determine variable 
-    //   start on propertyName
-    //   3. if node is already visited, return ""
-    //      visit node
-    //      retrieve custom property substitution val (with getCascadedPropertyValue)
-    //      *if substitution value is not a custom property, return it
-    //      get each individual variable of substitution val (with getIndividualVars)
-    //      for each variable 
-    //        go to node for custom property (get with ^var((--[\w-]+)
-    //          go to 3
-    //        if valid value wasnt found, get fallback with (var^(--[], (.*))
-    //        recurse to *
-    //          go to 3
-    //   Reset visits
-    const node = this._getNode(propertyName);
-    const substitute = node.resolveFor(this._element);
-    this._unvisitNodes();
-    return substitute;
-  }
-
-  contains(nodeName) {
-    return this._nodes.some(e => e.name === nodeName);
-  }
-
-  _createNode(nodeName) {
-    this._nodes.push({
-      name: nodeName,
-      outgoing: [],
-      visited: false,
-
-      find: function(nodeName) {
-        this.visited = true;
-        for (let outgoingNode of this.outgoing) {
-          if (outgoingNode.name === nodeName) {
-            return true;
-          }
-          if (!outgoingNode.visited) {
-            const wasFound = outgoingNode.find(nodeName);
-            if (wasFound) {
-              return wasFound;
-            }
-          }
-        }
-        return false;
-      },
-
-      resolveFor: function(element) {
-        if (this.visited) {
-          return guaranteedInvalid;
-        }
-        this.visited = true;
-
-        const substitute = getSpecifiedValue(element, this.name);
-        return makeCssVarSubstitutions(element, substitute);
-      },
-    });
-  }
-
-  _getNode(nodeName) {
-    return this._nodes.find(e => e.name === nodeName);
-  }
-
-  _drawEdgeConnecting(fromNodeName, toNodeName) {
-    if (!fromNodeName || !toNodeName) {
-      return;
-    }
-
-    const fromNodeRef = this._getNode(fromNodeName);
-    const toNodeRef = this._getNode(toNodeName);
-    if (!fromNodeRef) {
-      return;
-    }
-
-    if (!fromNodeRef.outgoing.some(e => e === toNodeRef)) {
-      fromNodeRef.outgoing.push(toNodeRef);
-    }
-  }
-
-  _attachNodes(fromNode, newNode) {
-    if (!isCustomProperty(newNode)) {
-      return;
-    }
-    
-    if (!this.contains(newNode)) {
-      this._createNode(newNode);
-      this._drawEdgeConnecting(fromNode, newNode);
-      const varSubstituteVal = getSpecifiedValue(this._element, newNode);
-      const dependencies = varSubstituteVal.value.matchAll(/var\((--[-\w]+)/g);
-
-      let outgoingEdges = [];
-      if (dependencies) {
-        outgoingEdges = [...dependencies].map(match => match[1]);
-      }
-      
-      for (let nodeAtEndOfEdge of outgoingEdges) {
-        this._attachNodes(newNode, nodeAtEndOfEdge);
-      }
-        
-    } else {
-      this._drawEdgeConnecting(fromNode, newNode);
-    }
-  }
-
-  _unvisitNodes() {
-    this._nodes.forEach(e => e.visited = false);    
-  }
-}
-
-// https://www.w3.org/TR/css-variables-1/#cycles
+// Implements two sections of the spec: 
+// [1] https://www.w3.org/TR/css-variables-1/#cycles
+// [2] https://www.w3.org/TR/css-variables-1/#using-variables
 function getVarSubstitute(element, property) {
+  // See [1] for cycles
   if (isCustomProperty(property)) {
-    let graph = new CustomPropertyGraph(element);
+    let graph = new CustomPropertyDependencyGraph(
+      property => getSpecifiedValue(element, property));
+    
     graph.createNodesFrom(property);
-
     if (graph.isCyclicAround(property)) {
       return guaranteedInvalid;
     }
 
-    return graph.findSubstitute(property);
-    
+    const substitute = getSpecifiedValue(element, property);
+    return makeCssVarSubstitutions(element, substitute);
+
+  // See [2] for using variables
   } else {
-    // If propertyName is not --a:
-    //   check if var() is in propertyInfo
-    //   if not:
-    //     return
-    //   if yes:
-    //     split into individual cssVar()
-    //     for each cssVar():
-    //       getVarSubstitute of customProp
-    //       if customProp does not evaluate:
-    //         getVarSubstitute of fallback
-    //     join results with ' '
     const cascadedProperty = getSpecifiedValue(element, property);
     let substitute = makeCssVarSubstitutions(element, cascadedProperty);
     if (!substitute.value) {
@@ -494,6 +253,7 @@ function getVarSubstitute(element, property) {
   }
 }
 
+// https://www.w3.org/TR/css-cascade-4/#defaulting-keywords
 function replaceKeywords(element, property, computedValue) {
   let resolvedValue = computedValue.value;
